+++ app/timer/page.tsx
@@ -1,11 +1,8 @@
 import { Counter } from '@/app/timer/components/Counter'
-import { Navigation } from '@/app/components/Navigation'
-import { routeMap } from '@/app/routes'
 
 export default function Timer() {
   return (
     <>
-      <Navigation currentPathname={routeMap.timer.path} />
       <Counter />
     </>
   )
